project/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
python-dotenv==1.0.1
redis==5.0.8
celery==5.4.0
chromadb==0.5.5
numpy==2.1.1
pypdf==5.0.1
youtube-transcript-api==0.6.2
beautifulsoup4==4.12.3
duckduckgo-search==6.3.0
slowapi==0.1.9
prometheus-client==0.20.0
structlog==24.4.0
email-validator==2.2.0
